refactor: moved JSON functions to storage.rs for modularity.
